refactor: move collections asserts to module
