+++ src/libs/collections/src/assert/rules.rs
