+++ src/libs/collections/src/assert/stores.rs
